Add new option --ignore-timestamps to osmium derive-changes
OSMObject::operator< does is used by derive-changes to determine if two
objects are equal. However, this operator does take the timestamp into
account. If the timestamp is missing on one object, it will be the
default (0) which leads to wrong and unexpected results.

Some users might prefert Osmium to produce a huge diff on the day,
personal metadata is not being published by the OSMF any more to clean
their own copies from personal information. On the other than, the diffs
will become very huge (larger than the input file because of
duplications and much larger if the input file was a PBF file).