Make sure there are no duplications due to attributes named like tags.
Attributes (such as id, version, timestamp, etc.) can appear in the
properties of the output with arbitrary configurable names. These could
overlap with tag keys which we don't want. This change removes tags with
those keys on the assumption that the names chosen for the attributes
are sufficiently different (something like "@id") from normal tag keys
that this will not happen very often and those tags are useless anyway.

Fixes #116.