Make calculation of CRC32 optional.
Calculating the CRC32 is very expensive and going without it makes the
program much much faster. So it is now optional. Use --crc/-c to enable
it. It will also be automatically enabled for JSON output to stay
compatible with earlier versions of Osmium which might be used in an
automated context (you can disable this with `--no-crc`). It is also
enabled if `-g data.crc32` is specified.