Bugfix: Allow index types with file name in export command
Index types dense_file_array and sparse_file_array are used by adding
the filename after a comma. This was not detected and those index types
were therefore not allowed.

This was already fixed in 1f5aea3afd103c35b5bba5880202379b13630fee for
the add-locations-to-ways command, but now is also fixed for the export
command.