Fix: Applying of change files from extracts
This is needed for a special case: When change files have been created
from extracts it is possible that they contain objects with the same
type, id, version, and timestamp. In that case we still want to get the
last object available. So we have to do the sorting correctly in that
case, too.