Add --keep-member-nodes option to add-locations-to-ways command
If this is set, all nodes referenced directly by relations in the input
file are kept in the output, even if they don't have any tags. This
adds some overhead, because the input file must be read twice and the
list of node ids must be kept in memory, but it is not huge.

See also #239.