Add options to the "diff" command to ignore certain attributes
When the new command line options --ignore-changeset, --ignore-uid, or
--ignore-user are used, the respective attributes of OSM objects are
ignored when comparing the input files. This is especially useful for
user names, because OSM users can change their own user names, which
happens often enough to matter.

When used with the OPL output format the ignored attributes are not
written out to the output.